+++ tests/unit/commands/test_base.py
@@ -37,19 +37,6 @@ def fake_command(app_metadata, fake_services):
     )
 
 
-def test_without_config(emitter):
-    """Test that a command can be initialised without a config.
-    This is pending deprecation but still supported.
-    """
-
-    with pytest.deprecated_call():
-        command = base.AppCommand(None)
-
-    emitter.assert_trace("Not completing command configuration")
-    assert not hasattr(command, "_app")
-    assert not hasattr(command, "_services")
-
-
 @pytest.mark.parametrize("always_load_project", [True, False])
 def test_needs_project(fake_command, always_load_project):
     """`needs_project()` defaults to `always_load_project`."""